Add files via upload
This was a project for Data 604 in which my group analyzed various factors affecting birth rates. We utilized datasets from the world bank.


The role and status of women around the world bears a direct impact on birth rates and demographic shifts. Female education, female labour force participation and contraception are some very important topics we will consider in our study on birth rates. Our overall data revolves around worldwide data for birth rates along with various other data sets relating to socio-economic conditions affecting females around the world. We believe this data is not just interesting to analyze but also can help us better understand the world around us. This data would be useful for the governments of the world to understand how birth rates can be affected by different socio-economic conditions, and this data may be helpful in understanding how to solve or mitigate population issues. We understand that data on birthrates is an issue that predominantly affects females, and therefore the research questions we came up with, aligned with our datasets will be more focused on females.